Moved Camera in seperate package.
