+++ ContextDesktop/src/test/java/com/github/migi_1/Context/model/TestEnvironment.java
@@ -15,6 +15,7 @@ import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
 
 import com.github.migi_1.Context.Main;
+import com.github.migi_1.Context.model.entity.Camera;
 import com.github.migi_1.Context.utility.ProjectAssetManager;
 import com.jme3.app.state.AppStateManager;
 import com.jme3.asset.AssetKey;
